commiting to

+++ website.bash
@@ -16,7 +16,7 @@ done
 
 # Build our search indexes now that we have HTML pages.
 echo "Building the search indexes"
-rm -fR pagefind
+if [ -d pagefind ]; then rm -fR pagefind; fi
 pagefind  --verbose --force-language en-US
 if [ -d .git ] ; then git add pagefind; fi
 echo "Done $(date)"
